scripts/requirements.txt
Pillow==10.4.0
tqdm==4.66.4